tester: create SaturnInventoryTest
