+++ noxfile.py
@@ -10,7 +10,7 @@ from typing import Sequence
 import nox
 from nox.sessions import Session
 
-nox.options.sessions = "lint", "mypy", "tests", "tests_pipeline"
+nox.options.sessions = "lint", "mypy", "tests", "example_tests"
 nox.options.reuse_existing_virtualenvs = True
 
 python_all_versions = ["3.9"]
@@ -58,10 +58,10 @@ def tests(session: Session) -> None:
 
 
 @nox.session(python=python_all_versions)
-def tests_pipeline(session: Session) -> None:
+def example_tests(session: Session) -> None:
     args = session.posargs
     install_project(session)
-    session.run("bash", "example/tests_pipeline", *args, external=True)
+    session.run("bash", "example/run_tests", *args, external=True)
 
 
 @nox.session(python=python_all_versions)
